job_search_workspace/applications/generator/requirements.txt
python-docx==1.1.2
pypdf==4.3.1

